+++ build-coordinator/src/test/java/org/jboss/pnc/coordinator/test/event/StatusUpdatesTest.java
@@ -39,6 +39,7 @@ import org.jboss.pnc.spi.BuildCoordinationStatus;
 import org.jboss.pnc.spi.BuildOptions;
 import org.jboss.pnc.spi.BuildResult;
 import org.jboss.pnc.spi.BuildSetStatus;
+import org.jboss.pnc.spi.RebuildMode;
 import org.jboss.pnc.spi.builddriver.BuildDriverResult;
 import org.jboss.pnc.spi.coordinator.BuildCoordinator;
 import org.jboss.pnc.spi.coordinator.BuildSetTask;
@@ -175,7 +176,7 @@ public class StatusUpdatesTest {
         AtomicInteger atomicInteger = new AtomicInteger(1);
 
         BuildOptions buildOptions = new BuildOptions();
-        buildOptions.setForceRebuild(true);
+        buildOptions.setRebuildMode(RebuildMode.FORCE);
         return buildTasksInitializer.createBuildSetTask(
                 buildConfigurationSet,
                 user,